requirements.txt
torch
torchvision
transformers
accelerate
bitsandbytes
timm
opencv-python
pandas
numpy
scikit-learn
matplotlib
tqdm
hydra-core
omegaconf
colorlog
